build: use flags for some feature

+++ app/build/util.ts
@@ -97,6 +97,7 @@ export async function crop({ path, tileset }: { path: string; tileset: string })
 export async function clean({path}:{path:string}) {
   if (!path)
     throw new RangeError("no path specified for cleaning!")
-  await Deno.remove(path, {recursive:true})
+  if (await Deno.lstat(path).then(() => true).catch(() => false))
+    await Deno.remove(path, {recursive:true})
   log.debug(`cleaned: ${path}`);
 }
